IGN: grr

+++ include/grace/thread.h
@@ -141,7 +141,7 @@ public:
 					 /// \return The unique id number.
 	unsigned int	 threadid (void)
 					 {
-					 	if (! tid) return (unsigned int) ((unsigned int) this & 0xffffffff);
+					 	if (! tid) return (unsigned int) ((unsigned long long) this & 0xffffffff);
 					 	return (unsigned int) ((unsigned long long) tid & 0xffffffff);
 					 }
